+++ database/migrations/2023_04_23_090615_create_rates_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
     {
         Schema::connection($this->connection)->create('rates', function (Blueprint $table) {
             $table->id();
-            $table->double('amount');
+            $table->float('amount');
             $table->string('updated_by');
             $table->timestamps();
         });
